Added ImportDiscordRoles component to isolate discord roles import
